MapGenerator/requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
tiktoken
pyperclip
python-dotenv
numpy<2.0,>=1.22.4
